refactor: folder structure for K8sPermissions
